02-RAG/rag_vertex_ai/pyproject.toml
[project]
name = "rag-vertex"
version = "0.1.0"
description = "Thin, evaluated wrapper around Vertex AI RAG memory service."
requires-python = ">=3.11"
dependencies = [
    "google-adk>=0.2",
    "google-api-core>=2.19",
    "google-genai>=0.8",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-cov>=5.0", "ruff>=0.4", "mypy>=1.10"]

[project.scripts]
rag-eval = "eval.harness:main"
rag-ask = "examples.answer_question:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_vertex", "examples", "eval"]

[tool.pytest.ini_options]
pythonpath = ["src", "."]
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.mypy]
strict = true
packages = ["rag_vertex"]
mypy_path = "src"
